requirements.txt
1.pandas
2.matplotlib
3.seaborn
4.scikit-learn
5.joblib
6.streamlit